+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 - Node.js: Fix internal serialization logic to handle `undefined` and `null` the same as `@pulumi/pulumi`
   (https://github.com/pulumi/pulumi-policy/pull/400).
 
+- Node.js: Add support for new metadata fields for policy packs and policies
+  (https://github.com/pulumi/pulumi-policy/pull/397).
+
 ---
 
 ## 1.17.0 (2025-07-29)